* implemented a new memory system which fixes a lot of memory leaks ** there is still one left in the cvar subsystem afaik * coding guidelines * removed pause functions (they are not really needed now that i think of it) * removed all Hunk_* functions * cleanup * cvar->value => integer checking * i think i did a lot more which i can't remember now ;-)

+++ src/game/game.h
@@ -223,7 +223,7 @@ typedef struct {
 	int (*GetModelAndName) (const char *team, character_t *chr);
 
 	/* managed memory allocation */
-	void *(*TagMalloc) (size_t size, int tag);
+	void *(*TagMalloc) (int size, int tag);
 	void (*TagFree) (void *block);
 	void (*FreeTags) (int tag);
